Added Walls
Pulled out camera zoom from 0.25 to 0.15 so that more of the map will be visible. Added the 'buildMap' function call in setup and corresponding declaration at the top of the file. Copied in 'buildMap' function from my version of the project. To make it work one new include was required, fstream. New global game object vector: MapObjects. Added texture files and map txt, as well as added them into the game's resident textures, and other necessary things to get these walls working.

It is not perfect and requires further adjustment, but it is functional.